+++ src/grids/icon/icon-analytical-model.ts
