+++ src/grids/index.ts
@@ -1,3 +1,3 @@
 export * from './interface';
 export * from './factory';
-export { IconGrid } from './icon';
+export { IconGrid } from './icon/icon';
