+++ src/index.ts
@@ -17,7 +17,7 @@ export { getColor, getColorScale } from './utils/styling';
 // Classes
 
 export { GridFactory, IconGrid } from './grids/index';
-export { IconGridAnalytical } from './grids/icon-analytical';
+export { IconGridAnalytical } from './grids/icon/icon-analytical';
 export { gridConfig } from './grids/factory';
 export type { IconGridMode } from './types';
